calculates sum of digits in a number

+++ MyPractice/PRACTICE_TOWARDS_PERFECTION/Logic_development/factorial_of_given_no.c
@@ -8,37 +8,37 @@
 
 typedef int status_t;
 
-int factorial_r(int n);
-int factorial(int n);
-int get_number();
+long long int factorial_r(long long int n);
+long long int factorial(long long int n);
+long long int get_number();
 
 int main(void){
-    int n, fact;
+    long long int n, fact;
     n = get_number();
     fact = factorial(n);
-    printf("factorial of %d using recurrsion: %d\n", n, fact);
+    printf("factorial of %lld using recurrsion: %lld\n", n, fact);
     fact = factorial_r(n);
-    printf("factorial of %d using loops: %d\n", n, fact);
+    printf("factorial of %lld using loops: %lld\n", n, fact);
 
     return (EXIT_SUCCESS);
 }
 
-int get_number(){
-    int n;
+long long int get_number(){
+    long long int n;
     printf("enter number: ");
-    scanf("%d", &n);
+    scanf("%lld", &n);
     return n;
 }
 
-int factorial(int n){
-    int i, fact = 1;
+long long int factorial(long long int n){
+    long long int i, fact = 1;
     for(i = n; i > 0; i--){
         fact = fact * i;
     }
     return fact;
 }
 
-int factorial_r(int n){
+long long int factorial_r(long long int n){
     if(n == 0){
         return 1;
     }
